Make Bits8-64 structs opaque to the programmer
Bonus: Fix the value for an argument not being evaluated if the
corresponding parameter is pseudo

+++ tests/xy_c_compile_resources/bitwiseOperations/bitBasics.c
@@ -10,5 +10,5 @@ uint32_t bitBasics_test(void) {
     const uint32_t e = a ^ b;
     const uint32_t f = d << 4 ^ e >> 5;
     const uint32_t g = ~(c | f & e);
-    return g;
+    return (uint32_t)g;
 }
